+++ web/src/PersonalContextPanel.test.tsx
@@ -99,6 +99,7 @@ function fakeApi(): Api {
     listAuditLogs: notImplemented,
     listAiConversations: notImplemented,
     getAiConversation: notImplemented,
+    listCaregiverMessages: notImplemented,
     listConversations: notImplemented,
     getConversation: notImplemented,
     listOperatorOrganizations: notImplemented,