+++ web/src/PreferencesPanel.test.tsx
@@ -95,6 +95,7 @@ function fakeApi(initial: PreferencePublic[]): {
     listAuditLogs: notImplemented,
     listAiConversations: notImplemented,
     getAiConversation: notImplemented,
+    listCaregiverMessages: notImplemented,
     listConversations: notImplemented,
     getConversation: notImplemented,
     listOperatorOrganizations: notImplemented,